+++ src/Views/Dashboard/Dashboard.test.tsx
@@ -26,7 +26,7 @@ describe("Dashboard", () => {
 
     const serverInfo: ServerInfo = {
         version: "0.4.0",
-        uptime: 1000n,
+        uptime: 1000n, // 16 minutes
         usage: 2000n,
         bucketCount: 2n,
         oldestRecord: 10n,
@@ -58,11 +58,11 @@ describe("Dashboard", () => {
     });
 
 
-    it("should show server info", async () => {
+    it("should show server info by default", async () => {
         client.getInfo = jest.fn().mockResolvedValue(serverInfo);
         client.getBucketList = jest.fn().mockResolvedValue([]);
 
-        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}}/>);
+        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}} />);
         const html = (await waitUntilFind(wrapper, "#ServerInfo")).hostNodes();
 
         expect(html.text()).toContain("0.4.0");
@@ -70,8 +70,16 @@ describe("Dashboard", () => {
         expect(html.text()).toContain("2 KB");
     });
 
+    it("should show license info when clicked on the tab", async () => {
+        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}} />);
+        const licenseTab = (await waitUntilFind(wrapper, ".ant-tabs-tab-btn")).hostNodes().at(1);
+        licenseTab.simulate("click");
+
+        expect(wrapper.text()).toContain("Business Source License (BUSL)");
+    });
+
     it("should show bucket info", async () => {
-        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}}/>);
+        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}} />);
         const bucket = (await waitUntilFind(wrapper, "#bucket_1")).hostNodes();
         expect(bucket.text()).toContain("bucket_1");
         expect(bucket.text()).toContain("1 KB");
@@ -79,7 +87,7 @@ describe("Dashboard", () => {
     });
 
     it("should order buckets by last records", async () => {
-        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}}/>);
+        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}} />);
         let bucket = (await waitUntilFind(wrapper, "#bucket_1"));
         expect(bucket.at(0).key()).toEqual("1");
 
@@ -88,10 +96,49 @@ describe("Dashboard", () => {
     });
 
     it("should push to BucketDetail if user click on bucket card", async () => {
-        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}}/>);
+        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}} />);
         const bucket = (await waitUntilFind(wrapper, "#bucket_1")).hostNodes();
         bucket.props().onClick();
 
         expect(mockPush).toBeCalledWith("/buckets/bucket_1");
+    });
+
+    it("shows warning icon if license is invalid", async () => {
+        client.getInfo = jest.fn().mockResolvedValue({
+            ...serverInfo,
+            license: {
+                licensee: "test",
+                plan: "free",
+                invoice: "123",
+                deviceNumber: "123",
+                expiryDate: Date.now() - 1000,
+                diskQuota: 1,
+                fingerprint: "123",
+            }
+        });
+
+        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}} />);
+        const warningIcon = (await waitUntilFind(wrapper, ".warningIcon")).hostNodes();
+        expect(warningIcon).toHaveLength(1);
+    });
+
+    it("does not show warning icon if license is valid", async () => {
+        client.getInfo = jest.fn().mockResolvedValue({
+            ...serverInfo,
+            license: {
+                licensee: "test",
+                plan: "free",
+                invoice: "123",
+                deviceNumber: "123",
+                expiryDate: Date.now() + 1000,
+                diskQuota: 1,
+                fingerprint: "123",
+            }
+        });
+
+        const wrapper = mount(<Dashboard backendApi={backend} permissions={{fullAccess: true}} />);
+        await new Promise(resolve => setTimeout(resolve, 100));
+        wrapper.update();
+        expect(wrapper.find(".warningIcon").hostNodes()).toHaveLength(0);
     });
 });